moves creation of the tabViewContainer into a Root class.
